pmcstat: fix sign-compare in multipart truncation guard
The truncated-record guard compares offset + len (both int) against
cc->pl_npc (uint32_t), which trips -Werror,-Wsign-compare in the
userland build (-Wsystem-headers is on there, unlike the kernel module
flags). offset and len are small non-negative values, so cast the sum
to uint32_t for the bounds check.

+++ usr.sbin/pmcstat/pmcstat_log.c
@@ -489,7 +489,7 @@ pmcstat_print_multipart(struct pmclog_ev_callchain *cc)
 			break;
 		} else if (type == PMC_CC_MULTIPART_CALLCHAIN) {
 			return (offset);
-		} else if (offset + len > cc->pl_npc) {
+		} else if ((uint32_t)(offset + len) > cc->pl_npc) {
 			/* Do not decode past the words the record carries. */
 			PMCSTAT_PRINT_ENTRY("truncated multipart record!");
 			break;
